+++ business_requirement_deliverable_cost/data/business_data.xml
@@ -4,10 +4,11 @@
 License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 -->
 <odoo noupdate="1">
-        <!-- Security Group Category -->
-        <record model="ir.module.category" id="module_category_business_requirement_cost">
-            <field name="name">Business Requirement Cost</field>
-            <field name="description"></field>
-            <field name="sequence">10</field>
-        </record>
+    <!-- Security Group Category -->
+    <record model="ir.module.category"
+            id="module_category_business_requirement_cost">
+        <field name="name">Business Requirement Cost</field>
+        <field name="description"></field>
+        <field name="sequence">10</field>
+    </record>
 </odoo>
