+++ business_requirement_deliverable_cost/models/__init__.py
@@ -1,1 +1,3 @@
 from . import business
+from . import business_requirement_deliverable
+from . import business_requirement_resource
