+++ business_requirement_deliverable_cost/models/business.py
@@ -4,236 +4,6 @@ from odoo import api, fields, models
 from odoo.addons import decimal_precision as dp
 
 
-class BusinessRequirementResource(models.Model):
-    _inherit = "business.requirement.resource"
-
-    sale_price_unit = fields.Float(
-        string='Sales Price'
-    )
-    sale_price_total = fields.Float(
-        compute='_compute_sale_price_total',
-        string='Total Revenue',
-        store=True
-    )
-    unit_price = fields.Float(
-        string='Cost Price'
-    )
-    price_total = fields.Float(
-        store=True,
-        compute='_compute_get_price_total',
-        string='Total Cost'
-    )
-    partner_id = fields.Many2one(
-        'res.partner',
-        related='business_requirement_deliverable_id.'
-        'business_requirement_id.partner_id',
-        string='Parter ID Related'
-    )
-
-    @api.multi
-    def _set_sales_price(self):
-        for resource in self:
-            pricelist_id = resource._get_pricelist()
-            resource.sale_price_unit = resource.product_id.lst_price
-            if pricelist_id and resource.partner_id \
-                    and resource.uom_id:
-                product = resource.product_id.with_context(
-                    lang=resource.partner_id.lang,
-                    partner=resource.partner_id.id,
-                    quantity=resource.qty,
-                    pricelist=pricelist_id.id,
-                    uom=resource.uom_id.id,
-                )
-                resource.sale_price_unit = product.price
-
-    @api.multi
-    def _set_cost_price(self):
-        for resource in self:
-            if not resource.product_id:
-                continue
-            qty_uom = 0
-            if resource.qty != 0:
-                qty_uom = resource.uom_id._compute_quantity(
-                    resource.qty,
-                    resource.product_id.uom_id
-                )
-            resource.unit_price = resource.product_id.standard_price * qty_uom
-
-    @api.multi
-    @api.depends('unit_price', 'qty')
-    def _compute_get_price_total(self):
-        for resource in self:
-            resource.price_total = resource.unit_price * resource.qty
-
-    @api.multi
-    @api.depends('sale_price_unit', 'qty')
-    def _compute_sale_price_total(self):
-        for resource in self:
-            resource.sale_price_total = resource.sale_price_unit * resource.qty
-
-    @api.multi
-    def _get_partner(self):
-        self.ensure_one()
-        br_id = br_deliverable = False
-        if self.business_requirement_deliverable_id.id:
-            br_deliverable = self.business_requirement_deliverable_id
-        if self.business_requirement_deliverable_id.\
-                business_requirement_id.partner_id:
-            br_id = br_deliverable.business_requirement_id
-        if br_id and br_id.partner_id:
-            return br_id.partner_id
-        else:
-            return False
-
-    @api.multi
-    def _get_pricelist(self):
-        self.ensure_one()
-        partner_id = self._get_partner()
-        if partner_id:
-            return (
-                partner_id.property_product_estimation_pricelist or
-                partner_id.property_product_pricelist)
-        return False
-
-    @api.multi
-    @api.onchange('product_id')
-    def product_id_change(self):
-        self.ensure_one()
-        super(BusinessRequirementResource, self).product_id_change()
-        self._set_sales_price()
-        self._set_cost_price()
-
-    @api.multi
-    @api.onchange('uom_id', 'qty')
-    def product_uom_change(self):
-        self.ensure_one()
-        # Calculte the sales_price_unit
-        self._set_sales_price()
-        # Calculate the unit_price
-        self._set_cost_price()
-
-
-class BusinessRequirementDeliverable(models.Model):
-    _inherit = "business.requirement.deliverable"
-
-    @api.multi
-    def _default_currency(self):
-        return self.env.user.company_id.currency_id
-
-    unit_price = fields.Float('Unit Price')
-    price_total = fields.Float('Total Price')
-    resource_task_total = fields.Float(
-        compute='_compute_resource_task_total',
-        string='Total Tasks',
-        store=True
-    )
-    resource_procurement_total = fields.Float(
-        compute='_compute_resource_procurement_total',
-        string='Total Procurement',
-        store=True
-    )
-    gross_profit = fields.Float(
-        string='Est. Gross Profit',
-        compute='_compute_gross_profit',
-        store=True
-    )
-    total_revenue_ci = fields.Float(
-        compute='_compute_total_revenue_in_ci',
-        string='Total Revenue',
-    )
-    company_currency_id = fields.Many2one(
-        comodel_name='res.currency',
-        string='Currency',
-        default=_default_currency,
-        help="Company Currency Id"
-    )
-    currency_status = fields.Boolean(
-        compute='_compute_currency_status',
-        string='Company Currency'
-    )
-
-    @api.depends('business_requirement_id.pricelist_id')
-    def _compute_currency_status(self):
-        for rec in self:
-            if rec.business_requirement_id and \
-                    rec.business_requirement_id.pricelist_id:
-                pricelist_id = rec.business_requirement_id.pricelist_id
-                if pricelist_id and pricelist_id.currency_id \
-                        and pricelist_id.currency_id.id == \
-                        rec.company_currency_id.id:
-                    rec.currency_status = True
-
-    @api.depends('currency_id')
-    def _compute_total_revenue_in_ci(self):
-        for rec in self:
-            total_revenue_ci = rec.currency_id._convert(
-                rec.price_total,
-                rec.company_currency_id,
-                rec.business_requirement_id.company_id,
-                fields.Date.today()
-            )
-            rec.total_revenue_ci = total_revenue_ci
-
-    @api.multi
-    @api.depends('resource_ids', 'resource_ids.price_total')
-    def _compute_resource_task_total(self):
-        for rec in self:
-            resource_task_total = sum(
-                rec.mapped('resource_ids').filtered(
-                    lambda r: r.resource_type == 'task').mapped(
-                    'price_total'))
-            rec.resource_task_total = rec.currency_id._convert(
-                resource_task_total,
-                rec.company_currency_id,
-                rec.business_requirement_id.company_id,
-                fields.Date.today()
-            )
-
-    @api.multi
-    @api.depends('resource_ids', 'resource_ids.price_total')
-    def _compute_resource_procurement_total(self):
-        for rec in self:
-            resource_procurement_total = sum(
-                rec.mapped('resource_ids').filtered(
-                    lambda r: r.resource_type == 'procurement').mapped(
-                    'price_total'))
-            rec.resource_procurement_total = rec.currency_id._convert(
-                resource_procurement_total,
-                rec.company_currency_id,
-                rec.business_requirement_id.company_id,
-                fields.Date.today()
-            )
-
-    @api.multi
-    @api.depends(
-        'price_total',
-        'resource_task_total',
-        'resource_procurement_total')
-    def _compute_gross_profit(self):
-        for rec in self:
-            gross_profit = rec.price_total - \
-                rec.resource_task_total - rec.resource_procurement_total
-            rec.gross_profit = rec.currency_id._convert(
-                gross_profit,
-                rec.company_currency_id,
-                rec.business_requirement_id.company_id,
-                fields.Date.today()
-            )
-
-    @api.multi
-    def action_button_update_estimation(self):
-        for deliverable in self:
-            if deliverable.resource_ids:
-                for resource in deliverable.resource_ids:
-                    resource._set_sales_price()
-                    resource._set_cost_price()
-
-    @api.multi
-    def action_button_update_total_revenue(self):
-        self.sale_price_unit = sum(self.resource_ids.
-                                   mapped('sale_price_total'))
-
-
 class BusinessRequirement(models.Model):
     _inherit = "business.requirement"
