+++ business_requirement_deliverable_cost/security/business_requirement_deliverable_security.xml
@@ -4,19 +4,22 @@
 License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 -->
 <odoo>
-        <record id="group_business_requirement_estimation" model="res.groups">
-            <field name="name">Business Requirement Estimation</field>
-            <field name="implied_ids" eval="[(4, ref('business_requirement.group_business_requirement_user'))]"/>
-            <field name="category_id" ref="business_requirement_deliverable_cost.module_category_business_requirement_cost"/>
-            <field name="users" eval="[(4, ref('base.user_root'))]"/>
-        </record>
+    <record id="group_business_requirement_estimation" model="res.groups">
+        <field name="name">Business Requirement Estimation</field>
+        <field name="implied_ids"
+               eval="[(4, ref('business_requirement.group_business_requirement_user'))]"/>
+        <field name="category_id"
+               ref="business_requirement_deliverable_cost.module_category_business_requirement_cost"/>
+        <field name="users" eval="[(4, ref('base.user_root')), (4, ref('base.user_admin'))]"/>
+    </record>
 
-        <record id="group_business_requirement_cost_control" model="res.groups">
-            <field name="name">Business Requirement Cost control</field>
-            <field name="implied_ids"
-                   eval="[(4, ref('business_requirement_deliverable_cost.group_business_requirement_estimation'))]"/>
-            <field name="category_id" ref="business_requirement_deliverable_cost.module_category_business_requirement_cost"/>
-            <field name="users" eval="[(4, ref('base.user_root'))]"/>
-        </record>
+    <record id="group_business_requirement_cost_control" model="res.groups">
+        <field name="name">Business Requirement Cost control</field>
+        <field name="implied_ids"
+               eval="[(4, ref('business_requirement_deliverable_cost.group_business_requirement_estimation'))]"/>
+        <field name="category_id"
+               ref="business_requirement_deliverable_cost.module_category_business_requirement_cost"/>
+        <field name="users" eval="[(4, ref('base.user_root')), (4, ref('base.user_admin'))]"/>
+    </record>
 
 </odoo>
